Added a program to generate a dataset from yfinance data
Program uses yfinance to gather data and processes the data to generate some of the more common market tickers I could find.  I also included a file that has a sample of the dataset that the program is generating.  I will have the program running and continue to generate more data until a sizeable amount of data is retrieved.